+++ src/vrm/gltf/extensions.rs
@@ -1,3 +1,4 @@
+pub mod vrmc_node_constraint;
 pub mod vrmc_spring_bone;
 pub mod vrmc_vrm;